[AIRFLOW-5203] Move GCP BigTable to core
This commit moves GCP BigTable from contrib to core.
For more information check AIP-21.
