+++ docs/howto/operator/gcp/bigtable.rst
@@ -24,7 +24,7 @@ Google Cloud Bigtable Operators
 
 All examples below rely on the following variables, which can be passed via environment variables.
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :start-after: [START howto_operator_gcp_bigtable_args]
     :end-before: [END howto_operator_gcp_bigtable_args]
@@ -34,7 +34,7 @@ All examples below rely on the following variables, which can be passed via envi
 BigtableInstanceCreateOperator
 ------------------------------
 
-Use the :class:`~airflow.contrib.operators.gcp_bigtable_operator.BigtableInstanceCreateOperator`
+Use the :class:`~airflow.gcp.operators.bigtable.BigtableInstanceCreateOperator`
 to create a Google Cloud Bigtable instance.
 
 If the Cloud Bigtable instance with the given ID exists, the operator does not compare its configuration
@@ -46,7 +46,7 @@ Using the operator
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_instance_create]
@@ -57,7 +57,7 @@ it will be retrieved from the GCP connection used. Both variants are shown:
 BigtableInstanceDeleteOperator
 ------------------------------
 
-Use the :class:`~airflow.contrib.operators.gcp_bigtable_operator.BigtableInstanceDeleteOperator`
+Use the :class:`~airflow.gcp.operators.bigtable.BigtableInstanceDeleteOperator`
 to delete a Google Cloud Bigtable instance.
 
 Using the operator
@@ -66,7 +66,7 @@ Using the operator
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_instance_delete]
@@ -77,7 +77,7 @@ it will be retrieved from the GCP connection used. Both variants are shown:
 BigtableClusterUpdateOperator
 -----------------------------
 
-Use the :class:`~airflow.contrib.operators.gcp_bigtable_operator.BigtableClusterUpdateOperator`
+Use the :class:`~airflow.gcp.operators.bigtable.BigtableClusterUpdateOperator`
 to modify number of nodes in a Cloud Bigtable cluster.
 
 Using the operator
@@ -86,7 +86,7 @@ Using the operator
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_cluster_update]
@@ -110,7 +110,7 @@ Using the operator
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_table_create]
@@ -129,7 +129,7 @@ Families <https://googleapis.github.io/google-cloud-python/latest/bigtable/colum
 BigtableTableDeleteOperator
 ---------------------------
 
-Use the :class:`~airflow.contrib.operators.gcp_bigtable_operator.BigtableTableDeleteOperator`
+Use the :class:`~airflow.gcp.operators.bigtable.BigtableTableDeleteOperator`
 to delete a table in Google Cloud Bigtable.
 
 Using the operator
@@ -138,7 +138,7 @@ Using the operator
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_table_delete]
@@ -152,7 +152,7 @@ BigtableTableWaitForReplicationSensor
 You can create the operator with or without project id. If project id is missing
 it will be retrieved from the GCP connection used. Both variants are shown:
 
-Use the :class:`~airflow.contrib.operators.gcp_bigtable_operator.BigtableTableWaitForReplicationSensor`
+Use the :class:`~airflow.gcp.sensors.bigtable.BigtableTableWaitForReplicationSensor`
 to wait for the table to replicate fully.
 
 The same arguments apply to this sensor as the BigtableTableCreateOperator_.
@@ -163,7 +163,7 @@ timeout hits and does not raise any exception.
 Using the operator
 """"""""""""""""""
 
-.. exampleinclude:: ../../../../airflow/contrib/example_dags/example_gcp_bigtable_operators.py
+.. exampleinclude:: ../../../../airflow/gcp/example_dags/example_bigtable.py
     :language: python
     :dedent: 4
     :start-after: [START howto_operator_gcp_bigtable_table_wait_for_replication]
