+++ tests/contrib/operators/postgres_local_executor.cfg
@@ -21,7 +21,7 @@
 # This configuration file is used to override
 # the executor to be LocalExecutor and switch to using Postgres as the database
 # Some system tests require LocalExecutor to be used because of parallelism required
-# One example is test_gcp_bigtable_operator_system.py in which sensor
+# One example is test_bigtable_system.py in which sensor
 # Has to be run in parallell with a task to create BigTable table
 [core]
 executor = LocalExecutor
