+++ tests/gcp/hooks/test_bigtable.py
@@ -29,7 +29,7 @@ from tests.contrib.utils.base_gcp_mock import mock_base_gcp_hook_no_default_proj
 from tests.compat import mock
 
 from airflow import AirflowException
-from airflow.contrib.hooks.gcp_bigtable_hook import BigtableHook
+from airflow.gcp.hooks.bigtable import BigtableHook
 
 CBT_INSTANCE = 'instance'
 CBT_CLUSTER = 'cluster'
@@ -44,7 +44,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
                         new=mock_base_gcp_hook_no_default_project_id):
             self.bigtable_hook_no_default_project_id = BigtableHook(gcp_conn_id='test')
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -56,7 +56,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -69,7 +69,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='example-project')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_instance_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -83,7 +83,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -98,7 +98,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         self.assertIsNone(res)
 
     @mock.patch('google.cloud.bigtable.instance.Instance.create')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_create_instance_missing_project_id(self, get_client, instance_create):
         operation = mock.Mock()
         operation.result_return_value = Instance(instance_id=CBT_INSTANCE, client=get_client)
@@ -114,7 +114,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         self.assertIn("The project id must be passed", str(err))
 
     @mock.patch('google.cloud.bigtable.instance.Instance.create')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_create_instance_overridden_project_id(self, get_client, instance_create):
         operation = mock.Mock()
         operation.result_return_value = Instance(instance_id=CBT_INSTANCE, client=get_client)
@@ -128,7 +128,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         instance_create.assert_called_once_with(clusters=mock.ANY)
         self.assertEqual(res.instance_id, 'instance')
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_table_missing_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -144,7 +144,7 @@ class TestBigtableHookNoDefaultProjectId(unittest.TestCase):
         err = cm.exception
         self.assertIn("The project id must be passed", str(err))
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_table_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -166,7 +166,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
                         new=mock_base_gcp_hook_default_project_id):
             self.bigtable_hook_default_project_id = BigtableHook(gcp_conn_id='test')
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -178,7 +178,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='example-project')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -191,7 +191,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='new-project')
         self.assertIsNotNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_instance_no_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -203,7 +203,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='example-project')
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -217,7 +217,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='example-project')
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_instance_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -231,7 +231,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='new-project')
         self.assertIsNone(res)
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_instance_no_instance(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -245,7 +245,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         get_client.assert_called_once_with(project_id='example-project')
 
     @mock.patch('google.cloud.bigtable.instance.Instance.create')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_create_instance(self, get_client, instance_create):
         operation = mock.Mock()
         operation.result_return_value = Instance(instance_id=CBT_INSTANCE, client=get_client)
@@ -259,7 +259,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         self.assertEqual(res.instance_id, 'instance')
 
     @mock.patch('google.cloud.bigtable.instance.Instance.create')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_create_instance_overridden_project_id(self, get_client, instance_create):
         operation = mock.Mock()
         operation.result_return_value = Instance(instance_id=CBT_INSTANCE, client=get_client)
@@ -273,7 +273,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         instance_create.assert_called_once_with(clusters=mock.ANY)
         self.assertEqual(res.instance_id, 'instance')
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_table(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -286,7 +286,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         instance_exists_method.assert_called_once_with()
         table_delete_method.assert_called_once_with()
 
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_delete_table_overridden_project_id(self, get_client):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -301,7 +301,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         table_delete_method.assert_called_once_with()
 
     @mock.patch('google.cloud.bigtable.table.Table.create')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_create_table(self, get_client, create):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -317,7 +317,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         create.assert_called_once_with([], {})
 
     @mock.patch('google.cloud.bigtable.cluster.Cluster.update')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_update_cluster(self, get_client, update):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -334,7 +334,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         update.assert_called_once_with()
 
     @mock.patch('google.cloud.bigtable.table.Table.list_column_families')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_list_column_families(self, get_client, list_column_families):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
@@ -350,7 +350,7 @@ class TestBigtableHookDefaultProjectId(unittest.TestCase):
         list_column_families.assert_called_once_with()
 
     @mock.patch('google.cloud.bigtable.table.Table.get_cluster_states')
-    @mock.patch('airflow.contrib.hooks.gcp_bigtable_hook.BigtableHook._get_client')
+    @mock.patch('airflow.gcp.hooks.bigtable.BigtableHook._get_client')
     def test_get_cluster_states(self, get_client, get_cluster_states):
         instance_method = get_client.return_value.instance
         instance_exists_method = instance_method.return_value.exists
